Add (optional) archived files metadata normalization
When archiving files from different machines, even if the source content is the same, different archive files may be generated  due to a variety of factors, like compression libraries or file metadata difference (permissions, etc.).

This creates problems in situations where it's expected that the source content should generate the same archive, if the content modified. For example, from a Git perspective, permissions 0644 and 0664 are the same, but they will cause different archives to be created on systems with different filesystems mount permission masks.

The `normalize_files_metadata` option tells the archiver to produce a normalized archive. In the Zip case, this is obtained by setting:

- the compression method to Store
- the modification date to a fixed one
- the permissions to 0644